requirements.txt
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.6.0
schedule>=1.2.0
python-dateutil>=2.8.2
Pillow>=10.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
python-docx>=0.8.11
newspaper3k>=0.2.8
